Fixup the MKL location logic
Since 2021, Intel provides a CMake package config file as part of the
oneAPI installs, but it is broken on Conda, so we can't fully switch to
it. Try to use the built-in script on systems that supports it (e.g.
non-Conda and non-Windows Conda), and use our fixed one otherwise.